requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyMuPDF==1.22.5  # Stable version that works with Python 3.12
pathlib>=1.0.1
tqdm>=4.65.0
click>=8.1.7
